requirements.txt
cachetools==4.0.0
fastapi==0.52.0
uvicorn==0.11.3
requests==2.23.0
motor==2.1.0
confluent-kafka==1.3.0
python-jose==3.1.0
python-dotenv==0.12.0
pytest
flake8